refactor(crane-mcp-remote): split tools.ts (513 LOC) into domain modules
- tools/shared.ts — TextContent, ToolResult types + staleWarning/textResult helpers
- tools/briefing.ts — crane_briefing section builders + registerBriefingTools
- tools/knowledge.ts — venture/doc/note/schedule/handoff/session handlers + registerDashboardTools + registerKnowledgeTools
- tools.ts — thin barrel calling all three register functions

All import sites unchanged. Verify clean. Gate: no files >500 LOC remain.
